lambda_response_handler/requirements.txt

psycopg2-binary==2.9.10
requests==2.31.0
